PATCH: Move internal `DbalSchemaDiff` and `DbalSchemaFactory` to `Neos\ContentRepository\Dbal`
This is just iternal utility code which is not used by the core itself but shared among dbal integrations for projections in example
