+++ Neos.ContentRepository.Core/composer.json
@@ -16,7 +16,6 @@
         "neos/error-messages": "*",
         "neos/utility-objecthandling": "*",
         "neos/utility-arrays": "*",
-        "doctrine/dbal": "^3.1.4",
         "symfony/serializer": "^6.3",
         "psr/clock": "^1",
         "behat/transliterator": "~1.0",
